Refactor into more organized form
